// components/numberField/numberFieldTargets.enum.ts
export enum NumberFieldTargets {
  root = 'root',
  errors = 'errors',
  helperText = 'helperText',
  input = 'input',
  inputPrefix = 'inputPrefix',
  inputSuffix = 'inputSuffix',
  label = 'label',
  readOnly = 'readOnly',
  tooltipButton = 'tooltipButton',
  tooltipPopover = 'tooltipPopover',
  simpleViewEdit = 'simpleViewEdit',
  simpleViewReadOnly = 'simpleViewReadOnly',
}